WG preview prepearing
