+++ packages/ui/src/pages/Profile/MyAccounts.tsx
@@ -5,12 +5,15 @@ import { PageHeader } from '../../components/page/PageHeader'
 import { PageTitle } from '../../components/page/PageTitle'
 import { SideBar } from '../../components/page/SideBar'
 import { Breadcrumbs } from '../../components/page/Sidebar/Breadcrumbs/Breadcrumbs'
+import { Statistics } from '../../components/statistics/Stats'
+import { useTotalBalances } from '../../hooks/useTotalBalances'
 import { Accounts } from './MyAccounts/Accounts'
-import { TotalBalances } from './MyAccounts/TotalBalances'
 import { MyProfileTabs } from './MyProfileTabs'
 import { MyProfile, MyProfileContent, ProfileSummary } from './Profile'
 
 export function MyAccounts() {
+  const { total, transferable, locked, recoverable } = useTotalBalances()
+
   return (
     <Page>
       <SideBar />
@@ -26,7 +29,14 @@ export function MyAccounts() {
             <PageTitle>My profile</PageTitle>
             <MyProfileTabs />
             <ProfileSummary>
-              <TotalBalances />
+              <Statistics
+                stats={[
+                  { title: 'Total balance', helperText: 'Lorem fishy', value: total },
+                  { title: 'Total transferable balance', helperText: 'Lorem fishy', value: transferable },
+                  { title: 'Total locked balance', helperText: 'Lorem fishy', value: locked },
+                  { title: 'Total recoverable', helperText: 'Lorem fishy', value: recoverable },
+                ]}
+              />
             </ProfileSummary>
           </PageHeader>
           <MyProfileContent>
