+++ packages/ui/src/pages/WorkingGroups/GroupsComponents.tsx
@@ -1,0 +1,9 @@
+import styled from 'styled-components'
+
+export const GroupsContainer = styled.div`
+  display: grid;
+  grid-template-columns: 1fr;
+  grid-template-rows: auto;
+  grid-row-gap: 24px;
+  width: 100%;
+`
